Update for VC++ 2010

+++ cl_dll/com_weapons.cpp
@@ -1,6 +1,6 @@
 /***
 *
-*	Copyright (c) 1996-2002, Valve LLC. All rights reserved.
+*	Copyright (c) 1996-2001, Valve LLC. All rights reserved.
 *	
 *	This product contains software technology licensed from Id 
 *	Software, Inc. ("Id Technology").  Id Technology (c) 1996 Id Software, Inc. 
@@ -273,5 +273,5 @@ stub functions for such things as precaching.  So we don't have to modify weapon
 int				stub_PrecacheModel		( char* s ) { return 0; }
 int				stub_PrecacheSound		( char* s ) { return 0; }
 unsigned short	stub_PrecacheEvent		( int type, const char *s ) { return 0; }
-const char		*stub_NameForFunction	( uint32 function ) { return "func"; }
+const char		*stub_NameForFunction	( unsigned long function ) { return "func"; }
 void			stub_SetModel			( edict_t *e, const char *m ) {}
